Add --pull-request flag
The --pull-request flag is optional and accepts the URL
of the PR, e.g. https://github.com/myorg/repo/pull/10

To make it easier to stay with protocol changes, this PR
now adds a submodule (to a private repo) that should only be fetched
when re-generation of the bindings is required.