Clean up output on error
Before this change, when an error occurred, the error would be printed
out, followed by the usage information, and then followed again by the
error message. Seeing the usage information tends to be misleading
since it traditionally suggests a problem parsing the program
arguments. This disables that behavior and removes the redundant error
message.