Revamp OCI image/architecture handling
When loading an SBOM from an OCI registry, if we encounter a multi-platform
image index we need a way to "resolve" that index to a single image. Currently
we just use the amd64 one, and error if one is not present.

This change has the loader take a configurable platform spec, and:

1. Filter available images based on that spec
2. Return the matching image

If there is not exactly one matching image we return an error.